Fixed Bugs
Fixed a bug in BBHEvolve where the step size was too big when numerically solving the Peters Equations. Fixed a small bug so now generate_invcdf correctly samples xvals from logarithmic space. Checked generate_invcdf and was able to successfully use it to sample from a standard normal distribution. I also found that the samples generated from generate_invcdf matched what we expect from the dtdfp distribution so long as we are careful to normalize appropriately.